backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
optional-django==0.3.0
djangorestframework-simplejwt==5.5.1
django-ordered-model==3.7.4
Pillow==12.0.0
django-nested-admin==4.1.6
django-friendship==1.9.6
drf-spectacular==4.12.2